fix: time out hung crontab installs
