+++ .changeset/crontab-timeout.md
@@ -1,0 +1,5 @@
+---
+"moadim": patch
+---
+
+Time out hung `crontab -` installs so daemon startup and routine sync report a clear error instead of wedging.
